Make devicestate sample and time variables unsigned

+++ src/wcmUSB.c
@@ -952,8 +952,8 @@ static int usbChooseChannel(WacomCommonPtr common, int device_type, unsigned int
 			}
 		}
 		DBG(1, common, "device with serial number: %u"
-		    " at %d: Exceeded channel count; ignoring the events.\n",
-		    serial, (int)wcmTimeInMillis());
+		    " at %u: Exceeded channel count; ignoring the events.\n",
+		    serial, wcmTimeInMillis());
 	}
 
 	return channel;
@@ -1334,7 +1334,7 @@ static void usbParseAbsEvent(WacomCommonPtr common,
 		change |= usbParseWacomAbsEvent(common, event, channel_number);
 	}
 
-	ds->time = (int)wcmTimeInMillis();
+	ds->time = wcmTimeInMillis();
 	channel->dirty |= change;
 }
 
@@ -1396,7 +1396,7 @@ static void usbParseAbsMTEvent(WacomCommonPtr common, struct input_event *event)
 			/* set this here as type for this channel doesn't get set in usbDispatchEvent() */
 			ds->device_type = TOUCH_ID;
 			ds->device_id = TOUCH_DEVICE_ID;
-			ds->sample = (int)wcmTimeInMillis();
+			ds->sample = wcmTimeInMillis();
 			break;
 
 		case ABS_MT_POSITION_X:
@@ -1415,7 +1415,7 @@ static void usbParseAbsMTEvent(WacomCommonPtr common, struct input_event *event)
 			change = 0;
 	}
 
-	ds->time = (int)wcmTimeInMillis();
+	ds->time = wcmTimeInMillis();
 	(&common->wcmChannel[private->wcmMTChannel])->dirty |= change;
 }
 
@@ -1506,7 +1506,7 @@ static void usbParseKeyEvent(WacomCommonPtr common,
 			/* time stamp for 2FGT gesture events */
 			if ((ds->proximity && !dslast->proximity) ||
 			    (!ds->proximity && dslast->proximity))
-				ds->sample = (int)wcmTimeInMillis();
+				ds->sample = wcmTimeInMillis();
 			break;
 
 		case BTN_TOOL_TRIPLETAP:
@@ -1518,7 +1518,7 @@ static void usbParseKeyEvent(WacomCommonPtr common,
 			/* time stamp for 2GT gesture events */
 			if ((ds->proximity && !dslast->proximity) ||
 			    (!ds->proximity && dslast->proximity))
-				ds->sample = (int)wcmTimeInMillis();
+				ds->sample = wcmTimeInMillis();
 			/* Second finger events will be considered in
 			 * combination with the first finger data */
 			break;
@@ -1527,7 +1527,7 @@ static void usbParseKeyEvent(WacomCommonPtr common,
 			change = 0;
 	}
 
-	ds->time = (int)wcmTimeInMillis();
+	ds->time = wcmTimeInMillis();
 	channel->dirty |= change;
 
 	if (change)
@@ -1558,7 +1558,7 @@ static void usbParseKeyEvent(WacomCommonPtr common,
 			change = 0;
 	}
 
-	ds->time = (int)wcmTimeInMillis();
+	ds->time = wcmTimeInMillis();
 	channel->dirty |= change;
 }
 
@@ -1620,7 +1620,7 @@ static void usbParseBTNEvent(WacomCommonPtr common,
 				change = 0;
 	}
 
-	ds->time = (int)wcmTimeInMillis();
+	ds->time = wcmTimeInMillis();
 	channel->dirty |= change;
 }
 
@@ -1906,7 +1906,7 @@ static void usbDispatchEvents(WacomDevicePtr priv)
 			if (event->code == REL_WHEEL)
 			{
 				ds->relwheel = -event->value;
-				ds->time = (int)wcmTimeInMillis();
+				ds->time = wcmTimeInMillis();
 				common->wcmChannel[channel].dirty |= TRUE;
 			}
 			else
